Restore changes from #21
The maven artifact fetch step introduced in #21 was intermittently failing due
to a bug in Java's initial implementation of TLS 1.3. The relevant RUN line in
the Dockerfile now forces Java to use TLS 1.2 which seems to fix the problem.
Upgrading the base image's conda openjdk package to something much newer
(or adding a RUN command to upgrade it here) should also fix the issue but that
would be a bit more of an invasive change.