Set scikit-image minimum version to 0.21.0
This avoids the problem introduced in 0.20.0 where the sample datasets are
downloaded to a cache directory upon import of the top-level module. This
causes an error when the container is being run as the calling user rather
than root as the default cache directory is not writable. It is also a waste
of time and network bandwidth as the datasets will never be used here.